+++ Source/Core/PureES.Tests/Generators/IncrementalGeneratorTests.Generate#PureES.CommandHandlers.intArrayCommandHandler.g.verified.cs
@@ -81,7 +81,7 @@ namespace PureES.CommandHandlers
         [global::System.ComponentModel.EditorBrowsableAttribute(global::System.ComponentModel.EditorBrowsableState.Never)]
         [global::System.Diagnostics.DebuggerStepThroughAttribute()]
         [global::System.Diagnostics.DebuggerNonUserCodeAttribute()]
-        public async global::System.Threading.Tasks.Task<ulong> Handle(int[] command, CancellationToken cancellationToken)
+        public async global::System.Threading.Tasks.Task<uint> Handle(int[] command, CancellationToken cancellationToken)
         {
             if (command == null)
             {
@@ -115,7 +115,7 @@ namespace PureES.CommandHandlers
                     }))
                 {
                     var result = global::PureES.Tests.Models.TestAggregate.CreateOnAsyncEnumerable(command, this._service0);
-                    var revision = ulong.MaxValue;
+                    var revision = uint.MaxValue;
                     if (result != null)
                     {
                         var events = new List<global::PureES.UncommittedEvent>();
