+++ Source/Core/PureES.Tests/Generators/IncrementalGeneratorTests.Generate.DotNet8_0#PureES.CommandHandlers.Dictionary_string_object_CommandHandler.g.verified.cs
@@ -78,7 +78,7 @@ namespace PureES.CommandHandlers
         [global::System.ComponentModel.EditorBrowsableAttribute(global::System.ComponentModel.EditorBrowsableState.Never)]
         [global::System.Diagnostics.DebuggerStepThroughAttribute()]
         [global::System.Diagnostics.DebuggerNonUserCodeAttribute()]
-        public async global::System.Threading.Tasks.Task<ulong> Handle(global::System.Collections.Generic.Dictionary<string, object> command, CancellationToken cancellationToken)
+        public async global::System.Threading.Tasks.Task<uint> Handle(global::System.Collections.Generic.Dictionary<string, object> command, CancellationToken cancellationToken)
         {
             if (command == null)
             {
