+++ Source/Core/PureES.Tests/Generators/IncrementalGeneratorTests.Generate.DotNet8_0#PureES.CommandHandlers.objectCommandHandler.g.verified.cs
@@ -78,7 +78,7 @@ namespace PureES.CommandHandlers
         [global::System.ComponentModel.EditorBrowsableAttribute(global::System.ComponentModel.EditorBrowsableState.Never)]
         [global::System.Diagnostics.DebuggerStepThroughAttribute()]
         [global::System.Diagnostics.DebuggerNonUserCodeAttribute()]
-        public async global::System.Threading.Tasks.Task<ulong> Handle(object command, CancellationToken cancellationToken)
+        public async global::System.Threading.Tasks.Task<uint> Handle(object command, CancellationToken cancellationToken)
         {
             if (command == null)
             {
@@ -112,7 +112,7 @@ namespace PureES.CommandHandlers
                     }))
                 {
                     var result = global::PureES.Tests.Models.ImplementedGenericAggregate.Create(command);
-                    var revision = ulong.MaxValue;
+                    var revision = uint.MaxValue;
                     if (result != null)
                     {
                         var e = new global::PureES.UncommittedEvent(result);
