+++ Source/Core/PureES/IConcurrency.cs
@@ -11,7 +11,7 @@ public interface IConcurrency
     /// <returns>
     /// The expected stream revision if available, otherwise <see langword="null" />
     /// </returns>
-    ulong? GetExpectedRevision(string streamId, object command);
+    uint? GetExpectedRevision(string streamId, object command);
     
     /// <summary>
     /// Sets the expected revision for a stream after a command is handled
@@ -20,5 +20,5 @@ public interface IConcurrency
     /// <param name="command">The handled command</param>
     /// <param name="previousRevision">The revision that the stream was at before the command was handled. <see langword="null"/> if stream did not exist.</param>
     /// <param name="currentRevision">The revision that the stream is currently at</param>
-    void OnUpdated(string streamId, object command, ulong? previousRevision, ulong currentRevision);
+    void OnUpdated(string streamId, object command, uint? previousRevision, uint currentRevision);
 }
