+++ Source/Core/PureES/RehydratedAggregate.cs
@@ -7,4 +7,4 @@
 /// <param name="StreamPosition">The stream position of the aggregate</param>
 /// <typeparam name="TAggregate">The Aggregate</typeparam>
 [PublicAPI]
-public record RehydratedAggregate<TAggregate>(TAggregate Aggregate, ulong StreamPosition);
+public record RehydratedAggregate<TAggregate>(TAggregate Aggregate, uint StreamPosition);
