+++ Source/Core/PureES/UncommittedEventsList.cs
@@ -3,10 +3,10 @@
 [PublicAPI]
 public class UncommittedEventsList
 {
-    public ulong? ExpectedRevision { get; }
+    public uint? ExpectedRevision { get; }
     public List<UncommittedEvent> Events { get; }
 
-    public UncommittedEventsList(ulong? expectedRevision, IEnumerable<object> @events)
+    public UncommittedEventsList(uint? expectedRevision, IEnumerable<object> @events)
     {
         ExpectedRevision = expectedRevision;
         Events = @events.Select(e => new UncommittedEvent(e)).ToList();
