+++ Source/EventStores/PureES.EventStore.CosmosDB/StreamPosition.cs
@@ -1,3 +1,3 @@
 ﻿namespace PureES.EventStore.CosmosDB;
 
-internal record StreamPosition(string EventStreamId, ulong Position);
+internal record StreamPosition(string EventStreamId, uint Position);
